+++ lib/redis/RedisGroupNotificationPublisher.ts
@@ -1,4 +1,4 @@
-import type { Redis } from 'ioredis'
+import type { RedisLike } from './RedisLike.js'
 import type { GroupNotificationPublisher } from '../notifications/GroupNotificationPublisher.js'
 import type { PublisherErrorHandler } from '../notifications/NotificationPublisher.js'
 import { DEFAULT_NOTIFICATION_ERROR_HANDLER } from '../notifications/NotificationPublisher.js'
@@ -25,10 +25,10 @@ export class RedisGroupNotificationPublisher<LoadedValue> implements GroupNotifi
   public readonly channel: string
   public readonly errorHandler: PublisherErrorHandler
 
-  private readonly redis: Redis
+  private readonly redis: RedisLike
   private readonly serverUuid: string
 
-  constructor(redis: Redis, config: RedisPublisherConfig) {
+  constructor(redis: RedisLike, config: RedisPublisherConfig) {
     this.redis = redis
     this.channel = config.channel
     this.serverUuid = config.serverUuid