+++ lib/redis/RedisNotificationConsumer.ts
@@ -1,4 +1,4 @@
-import type { Redis } from 'ioredis'
+import type { RedisLike } from './RedisLike.js'
 import { AbstractNotificationConsumer } from '../notifications/AbstractNotificationConsumer.js'
 import type { SynchronousCache } from '../types/SyncDataSources.js'
 import { defaultLogger } from '../util/Logger.js'
@@ -18,11 +18,11 @@ export class RedisNotificationConsumer<LoadedValue> extends AbstractNotification
   LoadedValue,
   SynchronousCache<LoadedValue>
 > {
-  private readonly redis: Redis
+  private readonly redis: RedisLike
   private readonly channel: string
   private messageHandler?: (channel: string, message: string) => void
 
-  constructor(redis: Redis, config: RedisConsumerConfig) {
+  constructor(redis: RedisLike, config: RedisConsumerConfig) {
     super(config.serverUuid)
     this.redis = redis
     this.channel = config.channel